refactor(866): make ITicketingBudgetService internal to Humans.Budget (ruling 43)
ITicketingBudgetService's only consumer, TicketingBudgetSyncJob, already
lives inside Humans.Budget, so the interface no longer needs to be public.
Moved it from Humans.Budget.Contracts into Humans.Budget/Services and
marked it internal. The sync job's constructor went internal too (a
public constructor can't take an internal parameter type), so its DI
registration moved from Shell into Humans.Budget's own Section.Register
via a factory.
