+++ src/Sections/Humans.Budget/Docs/Budget.md
@@ -230,7 +230,7 @@ Stored as string via `HasConversion<string>()`.
 **Owned tables:** `budget_years`, `budget_groups`, `budget_categories`, `budget_line_items`, `budget_audit_logs`, `ticketing_projections`
 **Status:** (A) Migrated (peterdrier/Humans PR for issue nobodies-collective/Humans#544, 2026-04-22). Moved into its own project `src/Sections/Humans.Budget` at G5 (nobodies-collective/Humans#866).
 
-- Everything but `Section`, `BudgetResource` and the migrations is `internal` — HUM0034 enforces it. The cross-section surface is the leaf project `Humans.Budget.Contracts`: `IBudgetServiceRead` (four methods), `ITicketingBudgetService` (one), `IBudgetDemoSeeder` (one), the DTOs those name, and the `BudgetYearStatus` / `ExpenditureType` enums. The leaf is a project rather than a `Contracts/` folder because one consumer is still in Base — `TicketQueryService`. `TicketingBudgetSyncJob` was the second until G5 lane 5b-3 moved it into this project's `Contracts/` folder.
+- Everything but `Section`, `BudgetResource` and the migrations is `internal` — HUM0034 enforces it. The cross-section surface is the leaf project `Humans.Budget.Contracts`: `IBudgetServiceRead` (four methods), `IBudgetDemoSeeder` (one), the DTOs those name, and the `BudgetYearStatus` / `ExpenditureType` enums. `ITicketingBudgetService` moved to `Humans.Budget.Services` and went internal (Peter's ruling 43) — its only consumer, `TicketingBudgetSyncJob`, is section-internal. The leaf is a project rather than a `Contracts/` folder because one consumer is still in Base — `TicketQueryService`. `TicketingBudgetSyncJob` was the second until G5 lane 5b-3 moved it into this project's `Contracts/` folder.
 - `BudgetService` lives in `Humans.Budget.Services` and depends only on Application-layer abstractions. `IBudgetService` (internal) is the full surface on top of the read interface; it stays an interface because the ticketing bridge's unit tests substitute it.
 - `BudgetRepository` (impl `src/Sections/Humans.Budget/Data/BudgetRepository.cs`, §15b Singleton + `IDbContextFactory<BudgetDbContext>`) is the only file that touches budget tables via `DbContext`. `IBudgetRepository` exposes atomic per-method operations — multi-entity mutations (e.g. creating a year with its default groups / categories / projection row, or syncing ticketing actuals + re-materializing projected line items) are single repository methods that do all their work inside one short-lived `DbContext`.
 - **Decorator decision — no caching decorator.** Budget is admin-only, low-traffic. Same rationale as Governance / User / Feedback.
