+++ src/Sections/Humans.Tickets/Docs/Tickets.md
@@ -226,7 +226,7 @@ have to see, **not** a compiler constraint: G5 lane 3c measured (2026-08-15,
 `dotnet msbuild <leaf>.csproj -t:ResolvePackageAssets -getItem:FrameworkReference`) that every
 leaf referencing `Humans.Interfaces` resolves `Microsoft.AspNetCore.App` transitively, so
 "the leaf stays framework-free" — this doc's earlier wording — is false and always was. All section services route every database
-read/write through `ITicketRepository`; neither `TicketQueryService.cs` nor `TicketSyncService.cs` imports `Microsoft.EntityFrameworkCore` or references `TicketsDbContext`. Umbrella issue nobodies-collective/Humans#545 closed by sub-tasks #545a (TicketQueryService → Application), #545b (TicketingBudgetService → Application, originally with its own `ITicketingBudgetRepository` — removed in #815), #545c (TicketSyncService → Application + `IShiftManagementService` / `IUserService` routing). `ITicketVendorService` connector split landed in peterdrier/Humans PR #277. The bridge itself has since left this section: with Budget's G5 move, `TicketingBudgetService` became internal to `Humans.Budget.Services` and its `ITicketingBudgetService` contract moved to `Humans.Budget.Contracts`, shrinking to `SyncActualsAsync` (projection refresh and parameter writes are Budget-admin-only, so they stayed off the published interface).
+read/write through `ITicketRepository`; neither `TicketQueryService.cs` nor `TicketSyncService.cs` imports `Microsoft.EntityFrameworkCore` or references `TicketsDbContext`. Umbrella issue nobodies-collective/Humans#545 closed by sub-tasks #545a (TicketQueryService → Application), #545b (TicketingBudgetService → Application, originally with its own `ITicketingBudgetRepository` — removed in #815), #545c (TicketSyncService → Application + `IShiftManagementService` / `IUserService` routing). `ITicketVendorService` connector split landed in peterdrier/Humans PR #277. The bridge itself has since left this section: with Budget's G5 move, `TicketingBudgetService` became internal to `Humans.Budget.Services` and its `ITicketingBudgetService` contract moved to `Humans.Budget.Services` and went internal (Peter's ruling 43), shrinking to `SyncActualsAsync` (projection refresh and parameter writes are Budget-admin-only, so they stayed off it).
 
 **Caching:** the §15 caching decorator pattern is applied (T-07, 2026-05-16). `CachingTicketQueryService` (`Services/Stores/`, Singleton — `ApplicationServicesTakeNoMemoryCacheRule` sweeps `Humans.*.Services` and would flag it in `Services/`) wraps a keyed-inner `TicketQueryService` (Scoped) and composes two tracked slices: `OrdersCache : TrackedCache<Guid, TicketOrderInfo>` (with `warmOnStartup: true`) keyed by ticket order id, and `UserHoldingsCache : TrackedCache<Guid, CachedUserTicketHoldings>` keyed by user id. The order slice owns the per-order `TicketOrderInfo` projection with attendees embedded; the user slice owns `UserTicketHoldings` and stores the 5-minute freshness deadline in the tracked value. The warm path calls the normal `ITicketService.GetTicketOrdersAsync` read method on the keyed inner, not a cache-named helper. Composition (multiple inner caches with different shapes) precludes inheriting `TrackedCache` directly, so the decorator implements `IHostedService` itself and forwards `StartAsync` to both tracked caches; the user slice has `warmOnStartup: false`. The only remaining `IMemoryCache` ticket entry owned by the decorator is per-event `TicketEventSummary` — whose key stays in Base's `CacheKeys` because `TicketTailorService` *populates* the same entry and the adapter must not name the section; the inner is cache-free and an architecture test pins no `IMemoryCache` in the inner's constructor.
